requirements.txt
google-api-python-client,
google-auth-oauthlib, 
python-dotenv, 
requests, 
beautifulsoup4, 
dateparser, 
google-generativeai